test(dashboard): ask the AST when a feature initialises, and stop letting the manifest hide one (#479, #480)
Review of the first commit found four ways past the checks it added. All four are closed here, each
with the reviewer's own mutation as the test.

1. "THE PAGE REALLY CALLS IT" ANSWERED THE WRONG QUESTION  [High]

   callsByName() means "reachable from somewhere", not "runs at load". Three shapes satisfied it
   while the feature never initialised:

     document.addEventListener("dfir-never", () => initSwimlane());
     function dead() { initSwimlane(); } void dead;
     el.onclick = () => initSwimlane();

   New helper loadTimeCallsTo() sits on walkLoadTime, which refuses to enter a function body that
   is not immediately invoked, so none of the three is visible to it. It returns every site rather
   than a boolean, and the check demands EXACTLY ONE — custody and the swimlane are not idempotent,
   and a second call stacks listeners and a ResizeObserver. Review's second-call mutation passed
   before; it fires now.

2. THE AUTO-RUN CHECK WAS STILL A REGEX  [High]

   `if (typeof document !== "undefined") (initCustodyButtons)();` self-runs in a real browser and
   matched neither pattern; `.call(...)` was a second spelling past them. The inverse was broken
   too: a COMMENT containing `(function initCustodyButtons() {})()` failed the suite for no reason.
   Now asked of the AST, with `(f)()`, `f.call()` and `f.apply()` collapsed to one answer.

3. THE MANIFEST COULD HIDE A WHOLE FEATURE  [High]

   Every check keys on FEATURES, so a feature MISSING from it is not examined at all — review
   deleted the Backup row, added a real module-load defect, and 115 tests passed. Comparing manifest
   names against names found in the page cannot fix that: both sides omit the same feature and
   agree.

   The expected set now comes from the PAGE'S SCRIPT TAGS, the one place a shipped feature cannot
   hide. Every dashboard-*.js it loads must be in FEATURES or in the new NON_FEATURES list, so a new
   module fails until someone classifies it.

   And ownership is checked, not just names: review moved initCustodyButtons's metadata onto the
   Anomalies row and all 124 tests passed, because the old check compared two deduplicated NAME
   sets. moduleGlobals() now has to agree that the module on that row really publishes it.

4. THE FILE WAS 871 LINES, PAST THE REPO'S 800 LIMIT  [Medium]

   check:size walks src/ and public/ only, so nothing failed — the cost was a file nobody reads the
   bottom of. Split into featureManifest.ts (shared data), dashboardFeatureModules.test.ts
   (structure), dashboardFeatureLifecycle.test.ts (when it runs, and what a 404 costs) and
   dashboardCensusSeam.test.ts. Largest is now 476.

FIVE MUTATIONS, each the reviewer's, each run against this commit:
  never-fired listener        -> "calls initSwimlane exactly once, at load"
  a second guarded call       -> same
  (initCustodyButtons)()      -> "does not auto-run initCustodyButtons on load"
  delete the Backup row       -> "classifies every dashboard module the page loads"
  move initializer metadata   -> "attributes every initializer to the module that publishes it"

The reachability version of the call check is deleted rather than kept alongside — two answers to
one question is the rot this file has now been bitten by three times.

Full suite 7,771 across 566 files. check:size, format:check, typecheck and lint clean.
