+++ companion/tests/dashboard/featureManifest.ts
@@ -1,0 +1,253 @@
+// The tier-3 feature manifest, shared by every suite that asks something per feature.
+//
+// Extracted from dashboardFeatureModules.test.ts when that file passed the repo's 800-line limit
+// (#415). The limit does not reach tests — check:size walks src/ and public/ only — so nothing
+// would have failed; growing the one file until nobody reads the bottom of it is the actual cost.
+
+import { readFile } from "node:fs/promises";
+import type { DashboardScript } from "../helpers/dashboardAst.js";
+import { dashboardScripts, functionBindingsOf, scriptFromSource } from "../helpers/dashboardAst.js";
+import { globalNamesOf } from "../helpers/dashboardModule.js";
+
+// TIER 3 (#415): whole features moved out of the inline script, each owning its own state.
+//
+// One suite for eight modules rather than eight suites, because the assertions are identical in
+// shape and the differences that matter are data — which names each publishes, which state each
+// keeps private. A per-module file would be eight copies of the same five checks.
+//
+// WHY THESE EIGHT, AND WHY AN IIFE. The measurement for this tier found nine features with ZERO
+// escaping reads: every mutable binding they touch is read by nothing else, so the binding travels
+// with the feature. That is the ADR's own plan for tier 3 — "they move into their feature's module
+// as `let` at module scope and never become anyone's API" — with one correction. In a CLASSIC
+// script a top-level `let` joins the shared global lexical environment, so it would still be
+// reachable by name from every other script; js/dashboard-tagger.js and js/dashboard-kev.js got
+// away with top-level declarations only because they hold no state. These do, so they are wrapped,
+// and only the names the page actually calls are published.
+
+export const DASHBOARD = new URL("../../../public/dashboard.html", import.meta.url);
+
+export interface Feature {
+  file: string;
+  /** Names the inline script calls by bare name, so the module must put them on `window`. */
+  publish: string[];
+  /** State that must NOT be reachable from outside the closure. */
+  private: string[];
+  /** A published entry point that does the feature's load-time work, if it has one. */
+  initializer?: string;
+  /** Names that only exist AFTER the initializer has run. */
+  postInitPublish?: string[];
+}
+
+export const FEATURES: Feature[] = [
+  {
+    file: "dashboard-anomalies.js",
+    publish: ["loadAnomalies", "scheduleAnomaliesReload", "markAnomalySpikeFalsePositive"],
+    private: ["anomaliesData", "anomaliesTimer"],
+  },
+  {
+    file: "dashboard-sessions.js",
+    publish: ["loadSessions", "scheduleSessionsReload", "summarizeSession"],
+    private: ["sessionsData", "sessionsTimer", "sessionSummaries"],
+  },
+  {
+    file: "dashboard-compliance.js",
+    publish: [
+      "loadCompliance",
+      "scheduleComplianceReload",
+      "setComplianceDiscovered",
+      "clearComplianceDiscovered",
+      "toggleComplianceFramework",
+    ],
+    private: ["complianceData", "complianceTimer"],
+  },
+  {
+    file: "dashboard-d3fend.js",
+    publish: ["loadD3fend", "scheduleD3fendReload"],
+    private: ["d3fendData", "d3fendTimer"],
+  },
+  {
+    file: "dashboard-geo.js",
+    publish: [
+      "loadGeoMap",
+      "scheduleGeoMapReload",
+      "renderGeoView",
+      "ensureGeoMap",
+      "renderGeoMarkers",
+      "geoFocusIp",
+      "geoDownloadCsv",
+    ],
+    private: [
+      "geoMapData",
+      "geoMap",
+      "geoLayer",
+      "geoFlowLayer",
+      "geoMapTimer",
+      "geoMapInitializing",
+      // Missing from the first inventory, which is its own argument for asserting the EXACT global
+      // set rather than listing names by hand and hoping the list is complete.
+      "geoTileUrl",
+    ],
+  },
+  {
+    file: "dashboard-custody.js",
+    publish: ["initCustodyButtons", "loadCustody", "verifyCustodyOnOpen"],
+    private: ["custodyRecords", "custodyFailedPaths", "custodyVerifiedAt"],
+    // Declared late. The page guards and calls initCustodyButtons() exactly as it does the other
+    // two initializers, but this entry omitted the field, so every check keyed on
+    // `f.initializer` skipped it — and the only thing standing behind it was
+    // `expect(html).toContain("initCustodyButtons();")`, a check on prose. The manifest
+    // completeness test below now derives the expected set from the page instead of trusting
+    // this list, so a fourth initializer cannot be added without an entry either.
+    initializer: "initCustodyButtons",
+  },
+  { file: "dashboard-backup.js", publish: ["loadCaseBackups", "restoreCaseBackup"], private: [] },
+  {
+    // The load-time-heavy one: everything it does on load is wrapped in initTicketIntegrations(),
+    // which the page calls where the block used to sit. openIrisImportModal is published from
+    // INSIDE that function, so it is not on this list — the exact-globals check below would
+    // otherwise fail, which is the honest signal that it appears later rather than at load.
+    file: "dashboard-tickets.js",
+    publish: ["pushFindingToTicket", "bulkPushFindingsToTicket", "initTicketIntegrations"],
+    private: [
+      // The re-entry latch this PR added. It was missing from the first version of this list, and
+      // moving it out to shared global scope passed every test in the file — the same lesson the
+      // geoTileUrl note above records, one tier later.
+      "initialised",
+      "pushSelect",
+      "notionHasDefault",
+      "clickupDefaultList",
+      "notionOverlay",
+      "irisImportOverlay",
+      "irisReconnectBtn",
+      "clickupOverlay",
+      "irisPushOverlay",
+    ],
+    // Everything this module does happens when the page calls initTicketIntegrations(), so the
+    // checks that matter have to RUN it — see the block at the bottom of this file.
+    initializer: "initTicketIntegrations",
+    postInitPublish: ["openIrisImportModal"],
+  },
+  {
+    file: "dashboard-collection-plan.js",
+    publish: ["fetchCollectionResults", "renderCollectionPlan"],
+    private: [],
+  },
+  {
+    // The canvas chart. Like dashboard-tickets.js, all of its load-time work is DOM wiring —
+    // eleven listeners on the canvas and toolbar plus a ResizeObserver — so it is wrapped in
+    // initSwimlane() and the page calls it where the old IIFE sat. Unlike tickets the initializer
+    // publishes nothing, so there is no postInitPublish list: all six names appear at load.
+    //
+    // swLocateInTable is NOT here on purpose. Its name says swimlane, its body scrolls a row in
+    // #forensicTimeline, and both callers are inside jumpToEvent, which stayed in the page.
+    file: "dashboard-swimlane.js",
+    publish: [
+      "loadSwimlane",
+      "scheduleSwimlaneReload",
+      "swRenderCanvas",
+      "swSelToolbar",
+      "swReflectSelection",
+      "initSwimlane",
+    ],
+    private: [
+      "SW_LANE_H",
+      "SW_AXIS_H",
+      "SW_DOT_R",
+      "SW_SEV_TOKEN",
+      "SW_LABEL_TOKEN",
+      "SW_AXIS_LABEL",
+      "swLanes",
+      "swDataMinMs",
+      "swDataMaxMs",
+      "swViewStartMs",
+      "swViewEndMs",
+      "swDrag",
+      "swDragMoved",
+      "swDragStartX",
+      "swDragViewStart",
+      "swHoverEvId",
+      "swSelEvId",
+      "swTimer",
+      "swRubber",
+      "swTimeBrush",
+    ],
+    // The thirteen private FUNCTIONS (swFitView, swUpdateSubtitle, swZoomRatio, swTsToX, swXToTs,
+    // swRenderLabels, swHitTest, swShowDetail, swUpdateZoomLabel, swSelectionChanged,
+    // swFinishRubber, swScopeToView, swExportPng) are deliberately not on that list: it exists to
+    // catch a binding that is assigned but never declared, and a function declaration cannot
+    // become an implicit global. That they stay off `window` is already asserted by the
+    // exact-globals check.
+    initializer: "initSwimlane",
+  },
+];
+
+/**
+ * The names a loaded module put on the global object, ignoring the sandbox's own furniture.
+ *
+ * The vm context is seeded with `window`/`globalThis`, the host globals the loader borrows live
+ * (Date, btoa, atob, console — see dashboardModule.ts) and whatever `extraGlobals` the caller
+ * supplied. None of those are the module's doing.
+ */
+const SANDBOX_FURNITURE = new Set(["window", "globalThis", "Date", "btoa", "atob", "console"]);
+const globalsOf = (api: Record<string, unknown>, seeded: string[] = []): string[] =>
+  globalNamesOf(api).filter((k) => !SANDBOX_FURNITURE.has(k) && !seeded.includes(k));
+
+export { SANDBOX_FURNITURE, globalsOf };
+
+export const read = (f: string) => readFile(new URL(`../../../public/js/${f}`, import.meta.url), "utf8");
+export const scripts = dashboardScripts();
+
+/**
+ * THE CENSUS SEAM: names the module binds that the inline script binds too.
+ *
+ * A named function rather than two lines inlined into the assertion, because the two lines are the
+ * thing that has now been wrong twice and neither wrongness was reachable from a test. Its own
+ * contract lives in the `describe` at the bottom of this file — synthetic module and inline sources
+ * carrying the exact mutations that got through, so the seam is exercised and not just its helper.
+ */
+const duplicateBindings = (
+  moduleName: string,
+  moduleSrc: string,
+  inlineScripts: DashboardScript[],
+): string[] => {
+  const declared = functionBindingsOf(scriptFromSource(moduleName, moduleSrc)).map((b) => b.name);
+  return inlineScripts.flatMap((s) =>
+    functionBindingsOf(s)
+      .filter((b) => declared.includes(b.name))
+      .map((b) => `${s.name}:${b.line} ${b.name}`),
+  );
+};
+
+export { duplicateBindings };
+
+/**
+ * The dashboard-*.js modules the page loads that are NOT tier-3 features.
+ *
+ * Exists so that "every module the page loads is classified" can be a real check. FEATURES alone
+ * cannot be the source of truth for what shipped: review deleted a row, added a genuine defect to
+ * that module, and the suite went green — the feature was simply not examined by anything. Listing
+ * the non-features explicitly means a new dashboard-*.js tag fails until someone decides which it
+ * is, rather than being skipped in silence.
+ *
+ * Nine pure helpers (#415 tier 1), four state owners (tier 2), the tagger and KEV panels — which
+ * are tier 3 but hold no state, so they have no private-binding contract to assert — and the no-op
+ * facade, which is tier 1 infrastructure.
+ */
+export const NON_FEATURES = new Set([
+  "dashboard-state.js",
+  "dashboard-escape.js",
+  "dashboard-time.js",
+  "dashboard-text.js",
+  "dashboard-glyphs.js",
+  "dashboard-filters.js",
+  "dashboard-ioc.js",
+  "dashboard-values.js",
+  "dashboard-fragments.js",
+  "dashboard-scope.js",
+  "dashboard-selection.js",
+  "dashboard-facets.js",
+  "dashboard-timeline-view.js",
+  "dashboard-tagger.js",
+  "dashboard-kev.js",
+  "dashboard-facade.js",
+]);
